server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.2
typing-extensions==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 